abstract send link form
