+++ frontend/src/pages/auth/ForgotPassword.tsx
@@ -1,10 +1,8 @@
-import { type FormEvent, useEffect, useState } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
+import { useState } from 'react';
 
-import { describeApiFailure, useForgotPassword } from '@/api';
-import { Button, Field, FieldLabel, Input } from '@/common/components';
+import { useForgotPassword } from '@/api';
 
-import { ErrorNote } from './ErrorNote';
+import { RequestLinkForm, SendLinkConfirmation } from './RequestLinkForm';
 import { WrapperWithLogo } from './Wrapper';
 
 /**
@@ -29,6 +27,21 @@ export const ForgotPassword = () => {
 
   const forgotPasswordMutation = useForgotPassword();
 
+  const handleSendLink = (submittedEmail: string) => {
+    forgotPasswordMutation.mutate(
+      { email: submittedEmail },
+      {
+        onSuccess: () => {
+          setStep('CONFIRMATION');
+        },
+      }
+    );
+  };
+
+  const handleResendLink = (submittedEmail: string) => {
+    forgotPasswordMutation.mutate({ email: submittedEmail });
+  };
+
   return (
     <WrapperWithLogo
       headerTitle={headerTitle}
@@ -36,177 +49,19 @@ export const ForgotPassword = () => {
       className="desktop:max-w-[362px] desktop:gap-8 gap-4 pt-35"
     >
       {step === 'FORM' ? (
-        <ForgotPasswordForm
+        <RequestLinkForm
           email={email}
           setEmail={setEmail}
-          mutation={forgotPasswordMutation}
-          onSuccess={() => setStep('CONFIRMATION')}
+          onSubmit={handleSendLink}
+          isPending={forgotPasswordMutation.isPending}
         />
       ) : (
-        <ResetLinkConfirmation
+        <SendLinkConfirmation
           email={email}
-          mutation={forgotPasswordMutation}
+          onResend={handleResendLink}
+          isPending={forgotPasswordMutation.isPending}
         />
       )}
     </WrapperWithLogo>
-  );
-};
-
-interface ForgotPasswordFormProps {
-  email: string;
-  setEmail: (email: string) => void;
-  mutation: ReturnType<typeof useForgotPassword>;
-  onSuccess: () => void;
-}
-
-const ForgotPasswordForm = ({
-  email,
-  setEmail,
-  mutation,
-  onSuccess,
-}: ForgotPasswordFormProps) => {
-  const [sendError, setSendError] = useState<string | null>(null);
-
-  const handleForgotPassword = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    setSendError(null);
-    mutation.mutate(
-      { email },
-      {
-        onSuccess: () => {
-          onSuccess();
-        },
-        onError: (error) => {
-          setSendError(sendFailureMessage(error));
-        },
-      }
-    );
-  };
-
-  return (
-    <>
-      <div>
-        {/* Form */}
-        <form onSubmit={handleForgotPassword} className="flex flex-col gap-6">
-          {/* Email Field */}
-          <Field>
-            <FieldLabel htmlFor="email">Email</FieldLabel>
-            <Input
-              id="email"
-              className="px-6"
-              type="email"
-              autoComplete="email"
-              placeholder="Enter your email"
-              value={email}
-              onChange={(e) => {
-                setEmail(e.target.value);
-                setSendError(null);
-              }}
-              required
-            />
-          </Field>
-
-          {sendError && <ErrorNote className="-mb-2">{sendError}</ErrorNote>}
-
-          {/* Send Link Button */}
-          <Button
-            type="submit"
-            variant="primary"
-            shape="default"
-            className="desktop:mt-6 mt-2 w-full py-3"
-            disabled={mutation.isPending}
-          >
-            Send link
-          </Button>
-        </form>
-
-        {/* Footer */}
-        <p className="desktop:mt-5 text-m-p2 tablet:font-medium tablet:mb-0 mt-6 mb-8 text-center">
-          <Link to="/login" className="text-blue-300 hover:underline">
-            Return to login
-          </Link>
-        </p>
-      </div>
-    </>
-  );
-};
-
-interface ResetLinkConfirmationProps {
-  email: string;
-  mutation: ReturnType<typeof useForgotPassword>;
-}
-
-const ResetLinkConfirmation = ({
-  email,
-  mutation,
-}: ResetLinkConfirmationProps) => {
-  const navigate = useNavigate();
-  const [countdown, setCountdown] = useState(60);
-  const [resendError, setResendError] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (countdown === 0) return;
-
-    const timer = setInterval(() => {
-      setCountdown((prev) => prev - 1);
-    }, 1000);
-
-    return () => clearInterval(timer);
-  }, [countdown]);
-
-  const handleResendClick = () => {
-    setResendError(null);
-    mutation.mutate(
-      { email },
-      {
-        onError: (error) => {
-          // Nothing was sent, so don't make them wait out a countdown for it.
-          setResendError(sendFailureMessage(error));
-          setCountdown(0);
-        },
-      }
-    );
-    setCountdown(60);
-  };
-
-  return (
-    <>
-      {/* Send Link Button */}
-      <div className="flex flex-col">
-        <Button
-          type="button"
-          variant="primary"
-          shape="default"
-          className="desktop:mt-2 w-full py-3"
-          onClick={() => navigate('/login')}
-        >
-          Return to login
-        </Button>
-        <p className="desktop:mt-2 text-m-p2 tablet:font-medium mt-3 py-3 text-center">
-          <button
-            onClick={
-              countdown > 0 || mutation.isPending
-                ? undefined
-                : handleResendClick
-            }
-            className={
-              countdown > 0 || mutation.isPending
-                ? 'cursor-not-allowed text-gray-400'
-                : 'cursor-pointer text-blue-300 hover:underline'
-            }
-          >
-            {countdown > 0
-              ? `Send again in ${countdown} seconds`
-              : 'Send link again'}
-          </button>
-        </p>
-        {resendError && (
-          <ErrorNote className="justify-center text-center">
-            {resendError}
-          </ErrorNote>
-        )}
-      </div>
-    </>
   );
 };
